requirements.txt
viam-sdk==0.25.2